Add pdf files and the Abschlussarbeit.tex to gitignore
Tracking changes in PDF files is not that useful and rquires a lot of 
space in the long run. Therefore those (binary) files should not be 
tracked if possible.
Tracking the changes in Abschlussarbeit.tex is also not that useful 
cause those are 100% redundant and can be easily reproduced by compiling 
the Abschlussarbeit.rmd file.